ewayshop/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
astroid==2.4.1
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
distlib==0.3.0
Django==3.1.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.20.2
pandas==1.2.4
Pillow==8.1.0
pyfcm==1.5.1
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2021.1
requests==2.25.1
s3transfer==0.3.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.0
urllib3==1.26.3
virtualenv==20.0.18
virtualenvwrapper-win==1.2.6
wrapt==1.12.1
